requirements.txt
azure-core==1.32.0
azure-storage-blob==12.24.0
caldav==1.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
debugpy==1.8.11
defusedxml==0.7.1
et_xmlfile==2.0.0
icalendar==6.1.0
idna==3.10
isodate==0.7.2
jira==3.8.0
msal==1.31.0
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
pycparser==2.22
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-gitlab==5.2.0
pywin32==308
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xlwings==0.33.5
